backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1
ultralytics>=8.2.0
torch>=2.3.0
torchvision>=0.18.0
opencv-python-headless>=4.9.0
imageio[ffmpeg]>=2.34.0
imageio-ffmpeg>=0.5.1
pydantic>=2.7.0
numpy>=1.26.0
sqlalchemy[asyncio]>=2.0
psycopg[binary,pool]>=3.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyyaml>=6.0
mlflow>=2.0
lap>=0.5.12
